CCF-transformation-numerique-competences-API/requirements.txt
careerjet-api==3.2
hydra-core
pymongo